+++ src/ua/edu/ukma/ykrukovska/unit5/homework/ParallelepipedFactory.java
@@ -2,11 +2,11 @@ package ua.edu.ukma.ykrukovska.unit5.homework;
 
 public class ParallelepipedFactory {
 
-    public static Parallelepiped createCube(int side) {
+    public static Parallelepiped createCube(double side) {
         return new Parallelepiped(side);
     }
 
-    public static Parallelepiped createPerfectPar(int base1Length, int base2Length, int height) {
+    public static Parallelepiped createPerfectPar(double base1Length, double base2Length, double height) {
         return new Parallelepiped(base1Length, base2Length, height);
     }
